fixing trefactoring the url_safety
